Fix multiple issues, causing no print and multispecie errors
Multiple issues were causing issue #9. The CRN printer was not up to
snuff, causing it to only print concentrations. Besides that, it did not
properly display to the user when an error occured. And the species
semantic action did not use the variable the map was moved to, which
also caused issues in reactions with multiple products or reactants