[Euler] Add support for mupltiple identical reactants (#15)
First, a bugfix. When a reaction like `x + x -> z` was parsed, when the
parser would try to add the second `x` to the species list, it would
replace the first `x`, and the reaction would end up being parsed as `x
-> z`, which is very wrong. That was fixed in the parser with the
`InsertToSpecieMap` function, which handles the whole thing a lot
better.

And the main focus of this commit was adding support for reactions of
the form `2x -> y`. This was added to the euler evaluator